app/requirements.txt
flask
flask-cors
pillow
google-generativeai
python-dotenv
requests
